requirements.txt
pydash==4.3.0
python-dateutil==2.6.1
tqdm==4.19.4
pySMART.smartx==0.3.7
pyudev==0.21.0
requests==2.18.4
requests-toolbelt==0.8.0
colorama
ereuse-utils[usb_flash_drive, session, cli]==0.3.0b7
click-spinner
click==6.7
inflection